Allow bar to configure metadata source.

+++ bar/app/AppModel.swift
@@ -2,12 +2,18 @@ import Foundation
 import Async
 import AsyncProcess
 import Awaitables
+import IDisposable
 
-public class AppModel
+public class AppModel : IDisposable
 {
 	var config: AmethystConfig? = nil
 	var contentServer: AsyncProcess? = nil
 	var metaServer: AsyncProcess? = nil
+
+	public func dispose() {
+		stopMetaServer()
+		stopContentServer()
+	}
 
 	public func loadConfig()
 	{
@@ -24,19 +30,41 @@ public class AppModel
 
 	public func startContentServer()
 	{
+		if (contentServer != nil) {
+			return
+		}
+
 		contentServer = config?.content.createProcess()
 		contentServer?.launch()
 	}
 
 	public func stopContentServer()
 	{
+		if (contentServer == nil) {
+			return
+		}
+
 		await (contentServer!.terminate())
 		contentServer = nil
 	}
 
 	public func startMetaServer()
 	{
+		if (metaServer != nil) {
+			return
+		}
+
 		metaServer = config?.meta.createProcess()
 		metaServer?.launch()
+	}
+
+	public func stopMetaServer()
+	{
+		if (metaServer == nil) {
+			return
+		}
+
+		await (metaServer!.terminate())
+		metaServer = nil
 	}
 }
